+++ src/components/flyout/flyout.tsx
@@ -1,23 +1,60 @@
-import React, { Component } from 'react';
+import React, { Component, CSSProperties } from 'react';
 import classnames from 'classnames';
-import PropTypes from 'prop-types';
 
 import { keyCodes, EuiWindowEvent } from '../../services';
 
+import { CommonProps, keysOf } from '../common';
 import { EuiFocusTrap } from '../focus_trap';
 import { EuiOverlayMask } from '../overlay_mask';
 import { EuiButtonIcon } from '../button';
 
-const sizeToClassNameMap = {
+type EuiFlyoutSize = 's' | 'm' | 'l';
+
+const sizeToClassNameMap: { [size in EuiFlyoutSize]: string } = {
   s: 'euiFlyout--small',
   m: 'euiFlyout--medium',
   l: 'euiFlyout--large',
 };
 
-export const SIZES = Object.keys(sizeToClassNameMap);
+export const SIZES = keysOf(sizeToClassNameMap);
 
-export class EuiFlyout extends Component {
-  onKeyDown = event => {
+export interface EuiFlyoutProps extends CommonProps {
+  onClose: () => void;
+  size?: EuiFlyoutSize;
+  /**
+   * Hides the default close button. You must provide another close button somewhere within the flyout.
+   */
+  hideCloseButton?: boolean;
+  /**
+   * Locks the mouse / keyboard focus to within the flyout
+   */
+  ownFocus?: boolean;
+  /**
+   * Specify an aria-label for the close button of the flyout.
+   */
+  closeButtonAriaLabel?: string;
+  /**
+   * Sets the max-width of the page,
+   * set to `true` to use the default size,
+   * set to `false` to not restrict the width,
+   * set to a number for a custom width in px,
+   * set to a string for a custom width in custom measurement.
+   */
+  maxWidth?: boolean | number | string;
+
+  style?: CSSProperties;
+}
+
+export class EuiFlyout extends Component<EuiFlyoutProps> {
+  static defaultProps: Partial<EuiFlyoutProps> = {
+    size: 'm',
+    hideCloseButton: false,
+    ownFocus: false,
+    closeButtonAriaLabel: 'Closes this dialog',
+    maxWidth: false,
+  };
+
+  onKeyDown = (event: KeyboardEvent) => {
     if (event.keyCode === keyCodes.ESCAPE) {
       event.preventDefault();
       this.props.onClose();
@@ -49,7 +86,7 @@ export class EuiFlyout extends Component {
 
     const classes = classnames(
       'euiFlyout',
-      sizeToClassNameMap[size],
+      sizeToClassNameMap[size!],
       widthClassName,
       className
     );
@@ -71,9 +108,6 @@ export class EuiFlyout extends Component {
     const flyoutContent = (
       <div
         role="dialog"
-        ref={node => {
-          this.flyout = node;
-        }}
         className={classes}
         tabIndex={0}
         style={newStyle || style}
@@ -101,42 +135,3 @@ export class EuiFlyout extends Component {
     );
   }
 }
-
-EuiFlyout.propTypes = {
-  className: PropTypes.string,
-  children: PropTypes.node,
-  onClose: PropTypes.func.isRequired,
-  size: PropTypes.oneOf(SIZES),
-  /**
-   * Hides the default close button. You must provide another close button somewhere within the flyout.
-   */
-  hideCloseButton: PropTypes.bool,
-  /**
-   * Locks the mouse / keyboard focus to within the flyout
-   */
-  ownFocus: PropTypes.bool,
-  /**
-   * Specify an aria-label for the close button of the flyout
-   */
-  closeButtonAriaLabel: PropTypes.string,
-  /**
-   * Sets the max-width of the page,
-   * set to `true` to use the default size,
-   * set to `false` to not restrict the width,
-   * set to a `number` for a custom width in `px`,
-   * set to a `string` for a custom width in a custom measurement.
-   */
-  maxWidth: PropTypes.oneOfType([
-    PropTypes.bool,
-    PropTypes.number,
-    PropTypes.string,
-  ]),
-};
-
-EuiFlyout.defaultProps = {
-  size: 'm',
-  hideCloseButton: false,
-  ownFocus: false,
-  closeButtonAriaLabel: 'Closes this dialog',
-  maxWidth: false,
-};