+++ src/components/flyout/index.ts
@@ -1,0 +1,7 @@
+export { EuiFlyout, EuiFlyoutProps, SIZES } from './flyout';
+
+export { EuiFlyoutBody, EuiFlyoutBodyProps } from './flyout_body';
+
+export { EuiFlyoutFooter, EuiFlyoutFooterProps } from './flyout_footer';
+
+export { EuiFlyoutHeader, EuiFlyoutHeaderProps } from './flyout_header';