+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed typo to correct `aria-modal` attribute in`EuiPopover` ([#2488](https://github.com/elastic/eui/pull/2488))
 - Fixed position of `EuiCodeBlock` controls and added more tests ([#2459](https://github.com/elastic/eui/pull/2459))
 - Changed `EuiCodeBlock` so that `overflowHeight` now applies a `maxHeight` instead of a `height` on the block ([#2487](https://github.com/elastic/eui/pull/2487))
+- Fixed potentially inconsistent state update ([#2481](https://github.com/elastic/eui/pull/2481))
 
 ## [`14.8.0`](https://github.com/elastic/eui/tree/v14.8.0)